requirements.txt
requests
pytest
allure-pytest
pylint
black
isort